+++ src/main/java/org/springframework/beans/factory/BeanFactory.java
@@ -1,8 +1,5 @@
 package org.springframework.beans.factory;
 
-import java.util.HashMap;
-import java.util.Map;
-
 /**
  * bean工厂：负责注册bean和获取bean
  *
@@ -10,15 +7,14 @@ import java.util.Map;
  * @date 2020/11/22
  * @project: sbs-small-spring
  */
-public class BeanFactory {
+public interface BeanFactory {
 
-  private final Map<String, Object> beanMap = new HashMap<>();
-
-  public void registerBean(String name, Object bean) {
-    beanMap.put(name, bean);
-  }
-
-  public Object getBean(String name) {
-    return beanMap.get(name);
-  }
+  /**
+   * 返回 Bean 的实例对象
+   *
+   * @param name 要检索的bean的名称
+   * @return 实例化的 Bean 对象
+   * @throws BeansException 不能获取 Bean 对象，则抛出异常
+   */
+  Object getBean(String name) throws BeansException;
 }