+++ src/main/java/org/springframework/beans/factory/support/DefaultSingletonBeanRegistry.java
@@ -1,0 +1,28 @@
+package org.springframework.beans.factory.support;
+
+import org.springframework.beans.factory.config.SingletonBeanRegistry;
+
+import java.util.HashMap;
+import java.util.Map;
+
+/**
+ * 单例Bean注册接口的默认实现
+ *
+ * @author: DoubleW2w
+ * @date: 2024/9/29
+ * @project: sbs-small-spring
+ */
+public class DefaultSingletonBeanRegistry implements SingletonBeanRegistry {
+
+  private final Map<String, Object> singletonObjects = new HashMap<>();
+
+  @Override
+  public Object getSingleton(String beanName) {
+    return singletonObjects.get(beanName);
+  }
+
+  @Override
+  public void registerSingleton(String beanName, Object singletonObject) {
+    singletonObjects.put(beanName, singletonObject);
+  }
+}