+++ src/test/java/org/springframework/beans/factory/HelloService.java
@@ -1,0 +1,13 @@
+package org.springframework.beans.factory;
+
+/**
+ * @author: DoubleW2w
+ * @date: 2024/9/29
+ * @project: sbs-small-spring
+ */
+public class HelloService {
+  public String sayHello() {
+    System.out.println("hello");
+    return "hello";
+  }
+}